% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_V2.R
\name{coin_V2}
\alias{coin_V2}
\title{A helper function to flip a random coin after a few rounds of fixed results. This prevents sequential computer introduced errors.}
\usage{
coin_V2(p, i, laged)
}
\arguments{
\item{p}{The probability of a head.}

\item{i}{The round ID code.}

\item{laged}{A window of time to check if the computer has already introduced an error.}
}
\description{
A helper function to flip a random coin after a few rounds of fixed results. This prevents sequential computer introduced errors.
}
