% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IPD_hmc.R
\name{fit_IPD_mcmc}
\alias{fit_IPD_mcmc}
\title{Fit a finite mixture model using optimization}
\usage{
fit_IPD_mcmc(
  d,
  covariates = NULL,
  n_strategies = 12,
  eta_moves = 4,
  eta_arb = 4,
  n_chains = 1,
  n_cores = 1,
  iterations = 1000,
  warmup = 500,
  adapt_delta = 0.95,
  max_treedepth = 12
)
}
\arguments{
\item{d}{A data list of the format exported by simulate_round_robin()}

\item{covariates}{A matrix of covariates. Row number should equal the number of players.}

\item{n_strategies}{The number of considered strategies. PrisonR considers 12 by default. Changing this requires advanced editing of the Stan code. Good luck!}

\item{eta_moves}{A term that controls the plausibility of implementation errors.}

\item{eta_arb}{A term that controls the plausibility of implementation errors.}

\item{n_chains}{Number of MCMC chains.}

\item{n_cores}{Number of cores to use.}

\item{iterations}{Total number of MCMC iterations.}

\item{warmup}{Warmup iterations: "warmup" must be < "iterations".}

\item{adapt_delta}{A term that controls MCMC proposal acceptance rate.}

\item{max_treedepth}{A term that controls the MCMC treedepth.}
}
\value{
A Stan object
}
\description{
Fit a finite mixture model using optimization
}
