% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IPD_optim.R
\name{fit_IPD_optim}
\alias{fit_IPD_optim}
\title{Fit a finite mixture model using optimization}
\usage{
fit_IPD_optim(
  d,
  covariates = NULL,
  n_strategies = 12,
  eta_moves = 4,
  eta_arb = 4
)
}
\arguments{
\item{d}{A data list of the format exported by simulate_round_robin().}

\item{covariates}{A matrix of covariates. Row number should equal the number of players.}

\item{n_strategies}{The number of considered strategies. PrisonR considers 12 by default. Changing this requires advanced editing of the Stan code. Good luck!}

\item{eta_moves}{A term that controls the plausibility of implementation errors.}

\item{eta_arb}{A term that controls the plausibility of implementation errors.}
}
\value{
A Stan object.
}
\description{
Fit a finite mixture model using optimization
}
