% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFTA.R
\name{GTFTA}
\alias{GTFTA}
\title{GTFTA function file.}
\usage{
GTFTA(
  i,
  coop,
  intent,
  coop_error,
  arb,
  arb_error,
  error_rate,
  arb_error_rate_type_1,
  arb_error_rate_type_2,
  standing,
  xi = 0,
  generosity_p = 0.2
)
}
\arguments{
\item{i}{The round ID code.}

\item{coop}{Was cooperation observed?}

\item{intent}{The the player intend to cooperate?}

\item{coop_error}{Was an error introduced by the computer?}

\item{arb}{Was the arbitrator called?}

\item{arb_error}{Did the arbitrator declare an error?}

\item{error_rate}{The rate at which the computer introduces errors of the form C to D.}

\item{arb_error_rate_type_1}{The arbitrators rate of failing to detect a real error.}

\item{arb_error_rate_type_2}{The arbitrators rate of claiming an error was a true defection.}

\item{standing}{A vector of standing.}

\item{xi}{The sneaky rate of the strategy.}
}
\value{
A vector of information.
\itemize{
  \item Cell 1 - Did the player cooperate?
  \item Cell 2 - Did the player intend tocooperate?
  \item Cell 3 - Did the computer introduce an error?
  \item Cell 4 - Did the player call the arbitrator?
  \item Cell 5 - Did the arbitrator declare an error occured?
  \item Cell 6-7 - Returned standing.
  \item Cell 8-9 - Mid-round standing.
}
}
\description{
GTFTA function file.
}
