% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_new_functions.R
\name{integrate_new_functions}
\alias{integrate_new_functions}
\title{A function to add new strategy files to the namespace}
\usage{
integrate_new_functions(path = path)
}
\arguments{
\item{path}{Full path to "PrisonersDilema"}
}
\description{
A function to add new strategy files to the namespace
}
