% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{A function to create the main directory used in the workflow}
\usage{
setup_folders(path = path, import_code = TRUE)
}
\arguments{
\item{path}{Full path to where folder will be stored. This folder will be called "PrisonersDilema"}
}
\description{
This function allows you to build the directories for the data, and to customize which folders appear
}
