% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sequence.R
\name{simulate_sequence}
\alias{simulate_sequence}
\title{Simulate a gameplay sequence between 2 strategies}
\usage{
simulate_sequence(
  n_rounds = 40,
  strategies = c("ATFT", "ATFT"),
  error_rate = 0.05,
  arb_error_rate_type_1 = 0.1,
  arb_error_rate_type_2 = 0.1,
  xi = c(0, 0),
  pID = c(1, 2)
)
}
\arguments{
\item{n_rounds}{The number of complete rounds to play}

\item{strategies}{A 2-vector of strategies to compete. Names must match legal function files: i.e., "ALLC", "TFT", "GTFT", "TFTA" etc.}

\item{error_rate}{The rate at which the computer introduces errors of the form C to D.}

\item{arb_error_rate_type_1}{The arbitrators rate of failing to detect a real error.}

\item{arb_error_rate_type_2}{The arbitrators rate of claiming an error was a true defection.}
}
\value{
A list of information.
\itemize{
  \item g_round - The game round ID code
  \item actor_id - Player ID codes
  \item arb - Was the arbitrator called to start the round?
  \item arb_err - Did the arbitrator declare an error occured?
  \item coop - Did an observed cooperation occur?
  \item coop_intent - Was cooperation intended?
  \item coop_err - Was an error in cooperation introduced by the computer.
  \item stand_1_Focal - End-round standing of focal.
  \item stand_1_Alter - End-round standing of alter.
  \item stand_mr_Focal - Mid-round standing of focal.
  \item stand_mr_Alter - Mid-round standing of alter.
}
}
\description{
Simulate a gameplay sequence between 2 strategies
}
