% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\name{updater}
\alias{updater}
\title{A helper function for ATFT}
\usage{
updater(tt, standing, f_is_g, coop, callArb, arbEr)
}
\arguments{
\item{tt}{The truth table for ATFT}

\item{standing}{The standing vector of focal and alter}

\item{f_is_g}{Is focal giver in this half round?}

\item{coop}{Was move cooperate or defect?}

\item{callArb}{Was arbitrator called?}

\item{arbEr}{Did arbitrator declare that an error occured?}
}
\description{
A helper function for ATFT
}
