% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_IPD_results.R
\name{visualize_IPD_results}
\alias{visualize_IPD_results}
\title{Plot a heatmap that shows the distribution of posterior probability over strategies for each player.}
\usage{
visualize_IPD_results(
  fit,
  d,
  mode = "optim",
  smart_sort = FALSE,
  color = "#440154FF",
  strategy_set = c("ALLD", "ALLC", "RANDY", "TFT", "TF2T", "GTFT", "WSLS", "TFTA",
    "TF2TA", "GTFTA", "WSLSA", "ATFT"),
  dashes_at = c(0)
)
}
\arguments{
\item{fit}{The results returned by either fit_IFD_optim or fit_IFD_mcmc.}

\item{d}{The data fed into either fit_IFD_optim or fit_IFD_mcmc.}

\item{mode}{Either 'optim' or 'mcmc'.}

\item{smart_sort}{Sort to cluster by strategy?}

\item{color}{The color of the 'hot' end of the heatmap.}

\item{strategy_set}{The list of strategies in the Stan model. The order here must match the order of strategies in the Stan file.}

\item{dashes_at}{A list of locations to plot dashed lines.}
}
\value{
A ggplot object. Heatmap.
}
\description{
Plot a heatmap that shows the distribution of posterior probability over strategies for each player.
}
