% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_models.R
\name{create_stan_models}
\alias{create_stan_models}
\title{A function to create the main directory used in the workflow}
\usage{
create_stan_models(path = path)
}
\arguments{
\item{path}{Full path to where files will be stored.}
}
\description{
This function allows you to build the directories for the data, and to customize which folders appear
}
